fix up dockerfile
This change modifies the dockerfile to specify the base image and move
where the files are installed. The rationale for moving the added files
is to leverage the default working dir and user for the base image. In
support of the file move, the app.py file has been updated for the new
csv data location.